style fixes and reorganize
