+++ package.json
@@ -5,9 +5,10 @@
   "main": "server.js",
   "scripts": {
     "test": "karma start karma.conf.js --single-run && npm run test_server",
-    "start": "npm run build && node server.js",
+    "start": "npm install && bower install && npm run build && node server.js",
     "dev": "webpack-dev-server --open --config webpack.dev.js",
     "build": "webpack --config webpack.prd.js",
+    "server": "node server.js",
     "test_server": "mocha test/unit/server/server_test.js --exit"
   },
   "keywords": [
@@ -28,7 +29,7 @@
     "babel-preset-react": "6.24.1",
     "chai": "4.1.2",
     "clean-webpack-plugin": "0.1.17",
-    "css-loader": "0.28.7",
+    "css-loader": "^0.28.7",
     "enzyme": "3.3.0",
     "enzyme-adapter-react-16": "1.1.1",
     "eslint": "4.16.0",
@@ -60,7 +61,7 @@
     "redux-thunk": "2.2.0",
     "sass-loader": "6.0.6",
     "sinon": "^4.2.2",
-    "style-loader": "0.19.0",
+    "style-loader": "^0.19.0",
     "supertest": "3.0.0",
     "uglifyjs-webpack-plugin": "0.4.6",
     "uuid": "3.1.0",
@@ -71,6 +72,7 @@
   "dependencies": {
     "babel-polyfill": "6.26.0",
     "body-parser": "1.18.2",
+    "bower": "^1.8.2",
     "express": "4.14.0"
   }
 }
