+++ server.js
@@ -3,7 +3,7 @@ const express = require('express');
 const path = require('path');
 const app = express();
 const bodyParser = require('body-parser');
-const task = require('./routes/task');
+const task = require('./server/routes/task');
 
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: true }));
