+++ server/routes/task.js
@@ -1,5 +1,5 @@
 const express = require('express');
-const tasksContainer = require('../tasks.json');
+const tasksContainer = require('../../tasks.json');
 const router = express.Router();
 
 /**
