+++ src/actions/types.js
@@ -1,3 +1,4 @@
 export const GET_TODOS = 'GET_TODOS';
 export const REQUEST_PENDING = 'GET_TODOS_PENDING';
 export const REQUEST_REJECTED = 'GET_TODOS_REJECTED';
+export const REQUEST_FULFILL = 'REQUEST_FULFILL';
