+++ src/components/Layout/styles.scss
@@ -1,12 +1,17 @@
+body {
+  font-family: 'Montserrat', sans-serif;
+  font-size: 16px;
+  margin: 0px;
+  padding: 0px;
+}
+
 .layout {
   background-image: url(./bg.jpg);
-  font-family: 'Montserrat', sans-serif;
   min-height: 768px;
 }
 
 .layout-container {
   max-width: 1200px;
   margin: 0 auto;
-  font-size: 16px;
   padding-top: 100px;
 }
