+++ src/index.html
@@ -6,7 +6,6 @@
         <title>Todo Task App</title>
         <meta name="description" content="">
         <meta name="viewport" content="width=device-width, initial-scale=1">
-        <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/normalize/7.0.0/normalize.min.css">
         <link href="https://fonts.googleapis.com/css?family=Montserrat:300,400,700" rel="stylesheet">
     </head>
     <body>
@@ -15,6 +14,5 @@
         <![endif]-->
 
         <div id="app"></div>
-        <script defer src="https://use.fontawesome.com/releases/v5.0.6/js/all.js"></script>
     </body>
 </html>
